+++ test-suite/TestAttachToSuspend/Program.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics;
 
 using DbgTest;
 using DbgTest.DAP;
@@ -10,23 +9,17 @@ namespace TestAttachToSuspend
 class Program
 {
     static void Main(string[] args)
-    {
+    {                                                                Label.Breakpoint("bp");
         Label.Checkpoint("init", "attach_test",
             (Object context) =>
             {
                 Context Context = (Context)context;
                 Context.Initialize(@"__FILE__:__LINE__");
                 Context.StartTargetAndAttach(@"__FILE__:__LINE__", StartSuspend: true);
-                Context.AddBreakpoint(@"__FILE__:__LINE__", "bp");
-                Context.SetBreakpoints(@"__FILE__:__LINE__");
+                Context.AddFunctionBreakpoint("Main");
+                Context.SetFunctionBreakpoints(@"__FILE__:__LINE__");
                 Context.ConfigurationDone(@"__FILE__:__LINE__");
             });
-
-        // wait some time, control process should attach and setup breakpoints
-        int i = 3000;
-        System.Threading.Thread.Sleep(3000);
-
-        i++;                                                        Label.Breakpoint("bp");
 
         Label.Checkpoint("attach_test", "finish",
             (Object context) =>